requirements.txt
Pillow>=10.0.0
requests>=2.28.0
click>=8.0.0
